+++ app/utils.r
@@ -16,9 +16,16 @@
 #' and should be removed or replaced when working with real spatial data.
 #'
 #' @param csv_path Character scalar. Path to a metadata CSV file on disk.
-#' @param deployment_data_path Character scalar. Path to a deployment_data CSV file on disk. Must follow the CMI template.
-#' @param rel_path_parts assumes the relative path captures important information that can be pared into its folder structure.  column names to be created by splitting the relative path on "/". Default = c("deployment", "region", "camera","survey_location")
-#' @param verbose 
+#' @param deployment_data_path Character scalar. Path to a deployment_data CSV
+#'   file on disk. Must follow the CMI template.
+#' @param rel_path_parts assumes the relative path captures important
+#'   information that can be pared into its folder structure.  column names to
+#'   be created by splitting the relative path on "/". Default = c("deployment",
+#'   "region", "camera","survey_location")
+#' @param verbose print informative messages
+#' @param site_name_cols names of columns to concatenate into the "site_name"
+#'   used for mapping.  These should make up the unique key identifying a
+#'   site/name
 #'
 #' @return A data.frame containing the cleaned and augmented metadata.
 #'   Additional columns include:
@@ -52,6 +59,7 @@
 load_metadata <- function(csv_path,
                           deployment_data_path,
                           rel_path_parts = c("deployment", "region", "camera","location_name"), 
+                          site_name_cols,
                           verbose = TRUE) {
     
     if (!file.exists(csv_path)) {
@@ -67,7 +75,7 @@ load_metadata <- function(csv_path,
         {if(!"camera"%in%names(.)){
             tidyr::separate_wider_delim(.,cols = relative_path,delim = "\\",names = rel_path_parts,cols_remove =FALSE)}
         }%>% 
-        mutate(camera=gsub(".*\\\\","",camera),site_name = camera, 
+        mutate(camera=gsub(".*\\\\","",camera),
                location_name =tolower(location_name),
                datetime=ymd_hms(date_time,tz = "UTC"))
     
@@ -77,7 +85,9 @@ load_metadata <- function(csv_path,
             janitor::clean_names() %>%
             mutate(deployment_datetime = mdy_hms(paste(deployment_date, deployment_time), tz = "UTC"),
                    retrieval_datetime = mdy_hms(paste(retrieval_date, retrieval_time), tz = "UTC"),
-                   location_name = tolower(location_name))
+                   location_name = tolower(location_name),
+                   site_name = paste(!!!syms(site_name_cols), sep = " - "))
+        
         meta <- meta %>%
             left_join(deploy,
                       by = join_by(location_name, region, camera == camera_name,