+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2026-08-11 - Allow site_name for mapping to be set per deployment PR #27
+
+* adds a configurable argument to the config that lets the user set the column names from which we build the site_name which is used in mapping to determine unique sites/points.
+
 # 2026-08-11 - Support dynamic fields in filter PR #22
 
 * Change hard coded field names in the field drop down filter to dynamically generated list based on all columns that have >1 value and do not have another filter already.